Add example sim, Fix example_P
Added the code for the half-normal example `example_sim`. Fixed an
issue with example_P where the output wasn't the correct dimension.
It turned out that the n_quad argument could lead to fewer quadrature
points than there are bins. To remedy this, I changed the defintion of
n_quad to be the number of quadrature points in the smallest bin.