Add ppdeconv, Rewrite derivs and loglik
ppdeconv now works when the data is in block diagonal format.
To implement this change, I needed to change some functions in derivs
and loglik so that the matrix algebra evaluated properly.
The output of some functions needed to be changed from sparse matrix
format to numeric vectors. Otherwise, optim would no work!