Add constructors
At this point in the project, I've realized that it would be best to
create S3 objects that govern the data and the fits from the models.
This is because I have to consider several different types of models
(mle, rmle, fixed P, variable P, etc) that use similar functions.
This commit represents the beginning of this change-over, where I've
defined a constructor for `ppdeconvObj` objects.