Change fit process to iterate over list
I think the best way to fit these models is to keep the data for
each block in a list and to iterate over it.  This is in contrast
to the previous method, where we tried to force the models into
one big block diagonal format. In that approach, it was harder to
define a function to update P.